+++ src/typing_ex/view/page.clj
@@ -7,7 +7,7 @@
    [taoensso.timbre :as timbre]
    [typing-ex.plot :refer [plot]]))
 
-(def ^:private version "1.3.10")
+(def ^:private version "1.4.4")
 
 (defn page [& contents]
   [::response/ok
@@ -54,49 +54,53 @@
        (filter #(= (:login %) login))
        count))
 
-(defn scores-page [ret user days ex-days]
+(defn- headline []
+ [:div.row
+     [:div.d-inline
+      [:a {:href "/" :class "btn btn-primary btn-sm"} "Go!"]
+      " "
+      [:a {:href "/sum/3" :class "btn btn-primary btn-sm"} "D.P."]
+      " "
+      #_[:a {:href "/daily" :class "btn btn-danger btn-sm"} "todays"]
+      " max "]
+     [:div.d-inline
+      (form-to [:get "/recent"]
+               (text-field {:size 2
+                            :value "7"
+                            :style "text-align:right"} "n"))]
+     [:div.d-inline
+      " days, "
+      [:a {:href "http://qa.melt.kyutech.ac.jp/"
+           :class "btn btn-info btn-sm"}
+       "QA"]
+      " "
+      [:a {:href "http://mt.melt.kyutech.ac.jp/"
+           :class "btn btn-info btn-sm"}
+       "MT"]
+      " "
+      [:a {:href "http://l22.melt.kyutech.ac.jp/"
+           :class "btn btn-info btn-sm"}
+       "L22"]
+      " "
+      [:a {:href "/logout" :class "btn btn-warning btn-sm"} "logout"]]])
+(defn scores-page [max-pt ex-days user days]
   ;;(timbre/debug ex-days)
   (page
-   [:h2 "Typing: Scores (last " days " days)"]
-   [:p
-    [:a {:href "/" :class "btn btn-primary btn-sm"} "Go!"]
-    " "
-    [:a {:href "/logout" :class "btn btn-warning btn-sm"} "logout"]
-    " "
-    [:a {:href "/trials" :class "btn btn-danger btn-sm"} "last40"]
-    " "
-    [:a {:href "/daily" :class "btn btn-danger btn-sm"} "todays"]
-
-    [:span {:class "mmm"} " "]
-    [:a {:href "http://qa.melt.kyutech.ac.jp/"
-         :class "btn btn-info btn-sm"}
-     "QA"]
-    " "
-    [:a {:href "http://mt.melt.kyutech.ac.jp/"
-         :class "btn btn-info btn-sm"}
-     "MT"]
-    " "
-    [:a {:href "http://l22.melt.kyutech.ac.jp/"
-         :class "btn btn-info btn-sm"}
-     "L22"]]
-
+   [:h2 "Typing: Last " days " days Maxes"]
+   [:p (headline)]
    [:p "直近 " days " 日間のスコア順リスト。カッコは通算練習日数。"]
    (into [:ol
-          (for [{:keys [max login]} ret]
+          (for [{:keys [max login]} max-pt]
             [:li
              max
-             "("
-             (count-ex-days ex-days login)
-             ") "
+             (format "(%d) " (count-ex-days ex-days login))
              [:a {:href (str "/record/" login)
                   :class (cond
                            (= login user) "yes"
                            :else "other")}
               login]])])
-   [:p
-    [:a {:href "/" :class "btn btn-primary btn-sm"} "Go!"]
-    " "
-    [:a {:href "/logout" :class "btn btn-warning btn-sm"} "logout"]]))
+   [:div (headline)]))   
+
 
 ;; not good
 (defn- ss [s]
@@ -109,7 +113,7 @@
         avg (/ (reduce + (map :pt (take 10 (reverse positives)))) 10.0)]
     (page
      [:h2 "Typing: " login " records"]
-     [:p "付け焼き刃はもろい。毎日、10分、練習しよう。"]
+     [:p "付け焼き刃はもろい。毎日、10 分、練習しよう。"]
      [:div (plot 300 150 positives)]
      [:br]
      [:ul
@@ -122,9 +126,8 @@
 (defn active-users-page [ret]
   (page
    [:h2 "Typing: Last 40 trials"]
-   [:p "最近の tp ユーザ 40 名。連続するセッションを１つとするが、
-        セッションの間に別ユーザが割り込むと別セッションとカウント。
-        改良するか？"]
+   [:p "最近の Typing ユーザ 40 件。連続するセッションを１つとカウントするが、
+        セッションの間に別ユーザが割り込むと別セッションに。改良するか？"]
    (into [:ol]
          (for [[u & _] ret]
            [:li (ss (:timestamp u)) " " (:login u)]))))
@@ -133,7 +136,19 @@
   ;;(timbre/debug ret)
   (page
    [:h2 "Typing: todays"]
-   [:p "本日の tp ユーザ。重複を省いて最終利用時間で並べ替え。"]
+   [:p "本日の Typing ユーザ。重複を省いて最終利用時間で並べ替え。"]
    (into [:ol]
          (for [r ret]
            [:li (ss (:timestamp r)) " " (:login r)]))))
+
+;; 自分は赤
+(defn sums-page [ret]
+  (page
+   [:h2 "Typing: Daily Points"]
+   [:p "直近の n 日タイピング平常点ポイント。上位の人、他の科目もな。"]
+   (into [:ol]
+         (for [r ret]
+           [:li (:sum r)
+            " "
+            [:a {:href (str "/record/" (:login r))}
+             (:login r)]]))))